[ci] Improvements for emulator test jobs
The setup execution time for non-device tests has been improved by
moving a handful of adb commands from BaseTest to DeviceTest setup.
These would run for nearly every test run and result in a lot of waiting
due to some RunProcess failures and timeouts.

The output from the shell command we run to check if a device is online
is now cached and only refreshed when required by certain tests/asserts.

The AssertHasDevices check has been consolidated into DeviceTest setup,
and removed from individual tests in most cases.

Attempts to restart the emulator if a test determines that it is
inaccessible have been fixed, and emulator data will not be reset in this
case.

I played around with a handful of emulator launch settings and found
that removing the no-boot-anim and headless UI options improved the
reliability of nightly LocalizationTests, and resulted in faster boot
times locally. These options are now disabled for those tests.

The DeploymentTest class has been replaced by TimeZoneInfoTests and
LocalizationTests, and the other tests in that class have been moved to
InstallAndRunTests. TimeZoneInfoTests and LocalizationTests will now
validate command line output rather than trying to press a button and
read a UI element. This should make them more reliable and faster.

The TimeZoneInfoTests and LocalizationTests suites have been moved into
separate test stages in the nightly test job.

All nightly tests now run against our .NET build/test artifacts rather
than classic XA.

An issue that resulted in LocalizationTests nodes 11 and 12 not
containing any tests has been fixed.

An issue that would cause LocalizationTests node 1 to run tests from
node 11 and node 12 has been fixed.

Test result attachments have been fixed for TimeZoneInfoTests and
LocalizationTests, ensuring that we capture the right logcat and build
files for each test variant.

An issue causing binlog files to be overwritten by tests that use
multiple build targets has been fixed by naming the binlog file after
the build log file.

Console output verbosity has been set to normal for dotnet test
invocations, which should help with debugging.

Test NUnit packages have been updated to their latest versions.
