+++ tests/MSBuildDeviceIntegration/Tests/InstallTests.cs
@@ -32,7 +32,6 @@ namespace Xamarin.Android.Build.Tests
 		public void ReInstallIfUserUninstalled ([Values (false, true)] bool isRelease)
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject () {
 				IsRelease = isRelease,
@@ -61,7 +60,6 @@ namespace Xamarin.Android.Build.Tests
 		public void InstallAndUnInstall ([Values (false, true)] bool isRelease)
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject () {
 				IsRelease = isRelease,
@@ -95,7 +93,6 @@ namespace Xamarin.Android.Build.Tests
 		public void ChangeKeystoreRedeploy ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject () {
 				PackageName = "com.xamarin.keytest"
@@ -125,7 +122,6 @@ namespace Xamarin.Android.Build.Tests
 		public void SwitchConfigurationsShouldRedeploy ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject () {
 				IsRelease = false,
@@ -171,7 +167,6 @@ namespace Xamarin.Android.Build.Tests
 		public void InstallWithoutSharedRuntime ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject () {
 				IsRelease = true,
@@ -229,7 +224,6 @@ namespace Xamarin.Android.Build.Tests
 		public void InstallErrorCode ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			//Setup a situation where we get INSTALL_FAILED_NO_MATCHING_ABIS
 			var abi = "armeabi-v7a";
@@ -252,7 +246,6 @@ namespace Xamarin.Android.Build.Tests
 		public void ToggleFastDev ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject {
 				EmbedAssembliesIntoApk = false,
@@ -339,7 +332,6 @@ namespace Xamarin.Android.Build.Tests
 		public void LoggingPropsShouldCreateOverrideDirForRelease ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var proj = new XamarinAndroidApplicationProject {
 				IsRelease = true,
@@ -357,7 +349,7 @@ namespace Xamarin.Android.Build.Tests
 			using (var builder = CreateApkBuilder ()) {
 				Assert.IsTrue (builder.Install (proj), "Install should have succeeded.");
 				RunAdbCommand ("shell setprop debug.mono.log timing");
-				Assert.True (builder.RunTarget (proj, "_Run"), "Project should have run.");
+				RunProjectAndAssert (proj, builder);
 				var didLaunch = WaitForActivityToStart (proj.PackageName, "MainActivity", Path.Combine (Root, builder.ProjectDirectory, "logcat.log"), 30);
 				RunAdbCommand ("shell setprop debug.mono.log \"\"");
 				Assert.True (didLaunch, "Activity should have started.");
@@ -371,7 +363,6 @@ namespace Xamarin.Android.Build.Tests
 		public void BlankAdbTarget ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var serial = GetAttachedDeviceSerial ();
 			var proj = new XamarinAndroidApplicationProject () {
@@ -423,7 +414,6 @@ namespace Xamarin.Android.Build.Tests
 		[TestCaseSource (nameof (AndroidStoreKeyTests))]
 		public void TestAndroidStoreKey (bool useApkSigner, bool isRelease, string packageFormat, string androidKeyStore, string password, string expected, bool shouldInstall)
 		{
-			AssertHasDevices ();
 			if (DeviceSdkVersion >= 30 && !useApkSigner && packageFormat == "apk") {
 				Assert.Ignore ($"Test Skipped. jarsigner and {packageFormat} does not work with API 30 and above");
 				return;
@@ -490,7 +480,6 @@ namespace Xamarin.Android.Build.Tests
 		public void LocalizedAssemblies_ShouldBeFastDeployed ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var path = Path.Combine ("temp", TestName);
 			var lib = new XamarinAndroidLibraryProject {
@@ -529,7 +518,6 @@ namespace Xamarin.Android.Build.Tests
 		public void IncrementalFastDeployment ()
 		{
 			AssertCommercialBuild ();
-			AssertHasDevices ();
 
 			var class1src = new BuildItem.Source ("Class1.cs") {
 				TextContent = () => "namespace Library1 { public class Class1 { public static int foo = 0; } }"
@@ -622,8 +610,6 @@ namespace Xamarin.Android.Build.Tests
 		[Test]
 		public void AdbTargetChangesAppBundle ()
 		{
-			AssertHasDevices ();
-
 			var proj = new XamarinAndroidApplicationProject {
 				IsRelease = true
 			};
